(breaking change) Move task runners out of task.py
The task runners have different responsibilities than the tasks. Their goal is
to call the tasks in a specific order. In this case, they should not be placed
directly in tasks.py.

Please note: this change introduce a breaking change from the previous
version.
