+++ example/test_flow.py
@@ -1,7 +1,7 @@
 import boto.swf.layer2 as swf
 
 from garcon import activity
-from garcon import task
+from garcon import runner
 import logging
 import random
 
@@ -22,14 +22,14 @@ def activity_failure(activity, context):
 
 
 test_activity_1 = create(
-    name='activity_1',
-    tasks=task.SyncTasks(
+    name='o',
+    tasks=runner.Sync(
         lambda activity, context: logger.debug('activity_1')))
 
 test_activity_2 = create(
     name='activity_2',
     requires=[test_activity_1],
-    tasks=task.AsyncTasks(
+    tasks=runner.Async(
         lambda activity, context: logger.debug('activity_2_task_1'),
         lambda activity, context: logger.debug('activity_2_task_2')))
 
@@ -37,10 +37,10 @@ test_activity_3 = create(
     name='activity_3',
     retry=10,
     requires=[test_activity_1],
-    tasks=task.SyncTasks(activity_failure))
+    tasks=runner.Sync(activity_failure))
 
 test_activity_4 = create(
     name='activity_4',
     requires=[test_activity_3, test_activity_2],
-    tasks=task.SyncTasks(
+    tasks=runner.Sync(
         lambda activity, context: logger.debug('activity_4')))
